게임 레벨 디자인 수정

+++ Assets/Script/UI/DefeatUI.cs
@@ -10,7 +10,7 @@ public class DefeatUIData : BaseUIData
     public string maxWaveText;
     public string currentWaveText;
 
-    public int defeatEarnedWisdomCoefficient;
+    public float defeatEarnedWisdomCoefficient;
     public int effectCount;
     public int effectDelay;
 }
